refactor: move FeatureTransformations class into build_counterfactual.py
