+++ chap_core/cli_endpoints/build_counterfactual.py
@@ -2,6 +2,7 @@
 
 from __future__ import annotations
 
+import ast
 import logging
 from pathlib import Path  # noqa: TC003 — used at runtime via cyclopts get_type_hints()
 from typing import Annotated
@@ -9,9 +10,76 @@ from typing import Annotated
 import pandas as pd
 from cyclopts import Parameter
 
-from chap_core.cli_endpoints.causal import FeatureTransformations
-
 logger = logging.getLogger(__name__)
+
+
+class FeatureTransformations:
+    """Safe arithmetic expression parsing, validation, and evaluation for counterfactual transforms."""
+
+    def __new__(cls, *args, **kwargs):
+        raise TypeError(f"{cls.__name__} instances cannot be created")
+
+    _ALLOWED_FUNCS = frozenset({"abs", "round"})
+
+    _ALLOWED_NODE_TYPES = (
+        ast.Expression,
+        ast.BinOp,
+        ast.UnaryOp,
+        ast.Name,
+        ast.Constant,
+        ast.Call,
+        ast.Load,
+        ast.Add,
+        ast.Sub,
+        ast.Mult,
+        ast.Div,
+        ast.Pow,
+        ast.USub,
+        ast.UAdd,
+    )
+
+    @staticmethod
+    def validate_expression(expr: str) -> None:
+        """Raise ValueError if expr is not a safe arithmetic expression in x."""
+        try:
+            tree = ast.parse(expr, mode="eval")
+        except SyntaxError as e:
+            raise ValueError(f"Invalid expression '{expr}': {e}") from e
+
+        for node in ast.walk(tree):
+            if not isinstance(node, FeatureTransformations._ALLOWED_NODE_TYPES):
+                raise ValueError(f"Disallowed construct in expression '{expr}'")
+            if isinstance(node, ast.Name) and node.id not in FeatureTransformations._ALLOWED_FUNCS and node.id != "x":
+                raise ValueError(f"Disallowed name '{node.id}' in expression '{expr}'")
+            if isinstance(node, ast.Constant) and not isinstance(node.value, (int, float)):
+                raise ValueError(f"Non-numeric constant in expression '{expr}'")
+            if isinstance(node, ast.Call) and not (
+                isinstance(node.func, ast.Name) and node.func.id in FeatureTransformations._ALLOWED_FUNCS
+            ):
+                raise ValueError(f"Disallowed function call in expression '{expr}'")
+
+    @staticmethod
+    def parse_transformations(transformations: list[str]) -> list[tuple[str, str]]:
+        """Parse ['col=expr', ...] into [('col', 'expr'), ...].
+
+        Raises ValueError if any entry lacks an '=' separator.
+        """
+        result = []
+        for t in transformations:
+            if "=" not in t:
+                raise ValueError(f"Transformation '{t}' is not in 'column=expression' format")
+            col, expr = t.split("=", 1)
+            result.append((col, expr))
+        return result
+
+    @staticmethod
+    def apply_transformation(series: pd.Series, expr: str) -> pd.Series:
+        """Apply expr to each non-NaN value; leave NaN values unchanged."""
+        namespace = {"__builtins__": {}, "abs": abs, "round": round}
+        result = series.copy()
+        mask = ~series.isna()
+        result[mask] = series[mask].apply(lambda x: eval(expr, namespace, {"x": x}))
+        return result
 
 
 def build_counterfactual_cmd(
